fix: prevent provider from edit/delete pending reserved products

+++ app/Services/Items/ProductService.php
@@ -4,6 +4,7 @@ namespace App\Services\Items;
 
 use App\Models\User;
 use App\Models\Product;
+use App\Enums\OrderStatus;
 use App\Http\Traits\FileTrait;
 use Illuminate\Support\Facades\DB;
 use App\Http\Traits\ApiResponseTrait;
@@ -32,6 +33,11 @@ class ProductService
 
     public function update($data, Product $product)
     {
+        // Check if this service is under reservation
+        if ($product->reservations()->whereRelation('morphReservation', 'status', OrderStatus::PENDING->value)->exists()) {
+            return $this->returnWrong('This product is under reservation, you cant edit it!');
+        }
+
         if (isset($data['thumbnail'])) {
             $fileName = $this->uploadFile($data['thumbnail'], $product->attachment_path, $product->thumbnail);
             $data['thumbnail'] = $fileName;
@@ -44,6 +50,10 @@ class ProductService
     {
         DB::beginTransaction();
         try {
+            // Check if this service is under reservation
+            if ($product->reservations()->whereRelation('morphReservation', 'status', OrderStatus::PENDING->value)->exists()) {
+                return $this->returnWrong('This product is under reservation, you cant delete it!');
+            }
             $product->delete();
             $this->removeDirectory($product->attachment_path);
             DB::commit();
